+++ src/Nonstandard/Uuid.php
@@ -17,6 +17,7 @@ namespace Ramsey\Uuid\Nonstandard;
 use Ramsey\Uuid\Codec\CodecInterface;
 use Ramsey\Uuid\Converter\NumberConverterInterface;
 use Ramsey\Uuid\Converter\TimeConverterInterface;
+use Ramsey\Uuid\Math\CalculatorInterface;
 use Ramsey\Uuid\Uuid as Rfc4122Uuid;
 use Ramsey\Uuid\UuidInterface;
 
@@ -34,11 +35,13 @@ class Uuid extends Rfc4122Uuid implements UuidInterface
         array $fields,
         NumberConverterInterface $numberConverter,
         CodecInterface $codec,
-        TimeConverterInterface $timeConverter
+        TimeConverterInterface $timeConverter,
+        CalculatorInterface $calculator
     ) {
         $this->fields = new Fields((string) hex2bin(implode('', $fields)));
         $this->codec = $codec;
         $this->numberConverter = $numberConverter;
         $this->timeConverter = $timeConverter;
+        $this->calculator = $calculator;
     }
 }